+++ test/prover/LayerZeroProver.t.sol
@@ -512,6 +512,6 @@ contract LayerZeroProverTest is BaseTest {
                 )
             }
         }
-        return abi.encodePacked(uint96(chainId), packed);
+        return abi.encodePacked(uint64(chainId), packed);
     }
 }
